// cosmowcl/src/app/shared/models/api/cosmopratiche/caricamentoPratica.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * Pratiche
 * API per la gestione pratiche 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StatoCaricamentoPratica } from './statoCaricamentoPratica.d';
import { RiferimentoEnte } from './riferimentoEnte.d';
import { RiferimentoUtente } from './riferimentoUtente.d';


export interface CaricamentoPratica { 
    id?: number;
    nomeFile?: string;
    pathFile?: string;
    identificativoPratica?: string;
    statoCaricamentoPratica?: StatoCaricamentoPratica;
    idPratica?: number;
    descrizioneEvento?: string;
    errore?: string;
    ente?: RiferimentoEnte;
    utente?: RiferimentoUtente;
    hasElaborazioni?: boolean;
    elaborazioni?: Array<CaricamentoPratica>;
    dtInserimento?: string;
    dtUltimaModifica?: string;
    dtCancellazione?: string;
    visualizzaDettaglio?: boolean;
}